backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
PyMuPDF==1.24.10
python-docx==1.1.2
sentence-transformers==3.1.1
numpy==2.1.1
